Extract build download logic to action
